Ajout méthodes CRUD ContactFacade

+++ src/java/services/ContactFacadeREST.java
@@ -34,11 +34,26 @@ public class ContactFacadeREST {
 
     private EntityManager em = Persistence.createEntityManagerFactory("ProjetJEEPU").createEntityManager();
     EntityTransaction tx = em.getTransaction();
-
+    
+    // Creation d'un contact
     @PUT
     @Path("create")
+    public String create(@QueryParam("login") String login, @QueryParam("password") String password) {
+        try {
+            Contact c = new Contact(1, login, password, Boolean.FALSE);
+            tx.begin();
+            em.persist(c);
+            tx.commit();
+            return "OK";
+        } catch(Exception e) {
+            return "Login déjà utilisé";
+        }
+    }
+
+    @PUT
+    @Path("createWithStatus")
     // exemple de creation (postman en put) : http://localhost:8080/ProjetJEE/webresources/contact/create?login=test&password=test&status=client
-    public String create(@QueryParam("login") String login, @QueryParam("password") String password, @QueryParam("status") String status) {
+    public String createWithStatus(@QueryParam("login") String login, @QueryParam("password") String password, @QueryParam("status") String status) {
         // Status : Client ou Admin
         // Si on met 1 l'id va s'autoincrement
         try {
@@ -86,13 +101,61 @@ public class ContactFacadeREST {
             tx.begin();
             Query q = em.createQuery("select c from Contact c where c.idcontact=:idparam");
             q.setParameter("idparam", id);
-            Contact c = (Contact) q.getSingleResult();
-            c.setLogin(login);
-            c.setPassword(password);
-            tx.commit();
-            return "OK";
+            if(q.getResultList().isEmpty()) {
+                return "Le contact n'existe pas / ID inconnu";
+            } else {
+                Contact c = (Contact) q.getSingleResult();
+                c.setLogin(login);
+                c.setPassword(password);
+                tx.commit();
+                return "OK";
+            }
         } catch (Exception e) {
-            return "Utilisateur inconnu";
+            return "Login déjà utilisé";
+        }
+    }
+    
+    // Modification du login
+    @POST
+    @Path("{id}/updateLogin")
+    public String updateLogin(@PathParam("id") Integer idcontact, @QueryParam("login") String login) {
+        try {
+            Query q = em.createQuery("select c from Contact c where c.idcontact=:idparam");
+            q.setParameter("idparam", idcontact);
+            if(q.getResultList().isEmpty()) {
+                return "Le contact n'existe pas / ID inconnu";
+            } else {
+                Contact c = (Contact) q.getSingleResult();
+                tx.begin();
+                c.setLogin(login);
+                em.persist(c);
+                tx.commit();
+                return "OK";
+            }
+        } catch(Exception e) {
+            return "Login déjà utilisé";
+        }
+    } 
+    
+    // Modification du password
+    @POST
+    @Path("{id}/updatePassword")
+    public String updatePassword(@PathParam("id") Integer idcontact, @QueryParam("password") String pass) {
+        try  {
+            Query q = em.createQuery("select c from Contact c where c.idcontact=:idparam");
+            q.setParameter("idparam", idcontact);
+            if(q.getResultList().isEmpty()) {
+                return "Le contact n'existe pas / ID inconnu";
+            } else {
+                Contact c = (Contact) q.getSingleResult();
+                tx.begin();
+                c.setPassword(pass);
+                em.persist(c);
+                tx.commit();
+                return "OK";
+            }
+        } catch(Exception e) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
         }
     }
     
@@ -101,14 +164,20 @@ public class ContactFacadeREST {
     // http://localhost:8080/ProjetJEE/webresources/contact/1/delete
     public String deleteContact(@PathParam("id") Integer id) {
         try {
-            tx.begin();
-            Query q = em.createQuery("delete from Contact c where c.idcontact=:idparam");
+            Query q = em.createQuery("select c from Contact c where c.idcontact=:idparam");
             q.setParameter("idparam", id);
-            q.executeUpdate();
-            tx.commit();
-            return "OK";
+            if(q.getResultList().isEmpty()) {
+                return "Le contact n'existe pas / ID inconnu";
+            } else {
+               tx.begin();
+                q = em.createQuery("delete from Contact c where c.idcontact=:idparam");
+                q.setParameter("idparam", id);
+                q.executeUpdate();
+                tx.commit();
+                return "OK"; 
+            }
         } catch (Exception e) {
-            return "Utilisateur inconnu";
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
         }
     }
     
@@ -120,7 +189,7 @@ public class ContactFacadeREST {
     public List<Contact> findAll() {
         try {
             Query q = em.createQuery("select c from Contact c");
-            return q.getResultList();
+            return q.getResultList(); 
         } catch(Exception e) {
             throw new WebApplicationException(Response.Status.BAD_REQUEST);
         }
@@ -130,40 +199,43 @@ public class ContactFacadeREST {
     // Modification du status du contact
     @POST
     @Path("{id}/updateState")
-    @Produces(MediaType.APPLICATION_JSON)
     // http://localhost:8080/ProjetJEE/webresources/contact/1/updateState?state=admin
     // TODO : renvoie "Unexpected 'E'"
     public String updateState(@PathParam("id") Integer idcontact, @QueryParam("state") String newstate)  {
         Query q = em.createQuery("select c from Contact c where c.idcontact=:idparam");
         q.setParameter("idparam", idcontact);
-        Contact client = (Contact) q.getSingleResult();
-        try {
-            tx.begin();
-            // On verifie quel etat doit changer
-            if(newstate.equals("admin")) {
-                if(!client.getAdmin()) {
-                    client.setAdmin(Boolean.TRUE);
-                    em.persist(client);
-                    tx.commit();
-                    return "OK";
+        if(q.getResultList().isEmpty()) {
+            return "Le contact n'existe pas / ID inconnu";
+        } else {
+            Contact client = (Contact) q.getSingleResult();
+            try {
+                tx.begin();
+                // On verifie quel etat doit changer
+                if(newstate.equals("admin")) {
+                    if(!client.getAdmin()) {
+                        client.setAdmin(Boolean.TRUE);
+                        em.persist(client);
+                        tx.commit();
+                        return "OK";
+                    } else {
+                        return "deja admin";
+                    }
+                } else if (newstate.equals("client")) {
+                    if(client.getAdmin()) {
+                        client.setAdmin(Boolean.FALSE);
+                        em.persist(client);
+                        tx.commit();
+                        return "OK";
+                    } else {
+                        return "deja client / team";
+                    } 
                 } else {
-                    return "deja admin";
-                }
-            } else if (newstate.equals("client")) {
-                if(client.getAdmin()) {
-                    client.setAdmin(Boolean.FALSE);
-                    em.persist(client);
-                    tx.commit();
-                    return "OK";
-                } else {
-                    return "deja client / team";
-                } 
-            } else {
-                return "Erreur dans la requete";
-            }    
-            
-        } catch(Exception e) {
-            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+                    return "Erreur dans la requete";
+                }    
+
+            } catch(Exception e) {
+                throw new WebApplicationException(Response.Status.BAD_REQUEST);
+            }
         }
     }
 }
